+++ CHANGES.md
@@ -2,6 +2,7 @@
 * improved transient saving to save optimized version of transient rather that whole API response
 * changed _Refresh Cache_ to POST to only run once.
 * fixed `API::wp_update_response` to properly reset the update transient after a shiny update
+* fixed Bitbucket authentication during AJAX update
 
 #### 6.0.0 / 2016-10-26
 * added `class Language_Pack` and new repo, [Language Pack Maker](https://github.com/afragen/github-updater-language-pack-maker), to create and update from a separate Language Pack repository.